Push the first iteration of the home page.
This contains a number of supporting components and refinements to the
build system. Unfortunately, I wouldn't call this a "clean" commit. It's
bloated, contains deps in the package.json that shouldn't be there, and
contains no tests. That said, using git is a learning process and Nate
could use some example code!